Add standard external_id/system columns with indexes
There is not unique index on the donation table as right now we have errors in
production db and there are recurring donations with changing value - these
should be split into many RC's. This means there will be duplicate external_ids
for such RC in donations.

Also change failure_count to fail_count as agreed.