Refactor keanu engine framework
New abstractions to accomodate for a need for many sources.

Batch -- a series of data transformig scripts
 `- has N >= 1 sources -- of DBSource or other type
 `- has 1 destination -- of DBDestination or other type
 `- has N >= 1 transforms -- of SQLTransform or other type
     `- each having N scripts -- single sql files (LoadScript class)
        `- each script having N statements

The Batch is defined in a yaml file. See `cli/keanu.yaml.sample`.
The SQL transform always runs in the destination db. This is why sources do not
have `url` param. They are "local" to destination.