Introduce 1-second minimum group join time
When someone joines and leaves a group-based segment (like Members) for 0
duration, we cannot express it with contact_segment. There are two solutions:
remove this 0-duration membership, or make it 1-second. The latter solution does
not wipe out the fact that someone was a member according to our data, so I'm
setting such 1-second join time in this case.